Revert "Delete qr code player"
