+++ .github/workflows/android.yml
@@ -2,9 +2,7 @@ name: Android CI
 
 on:
   push:
-
     branches: [ main, Staging-, Integrated ]
-
   pull_request:
     branches: [ main, Staging-, Integrated ]
