+++ .idea/misc.xml
@@ -3,7 +3,11 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="../../../../layout/custom_preview.xml" value="0.2677083333333333" />
+        <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/app/src/main/res/layout/activity_other_profile.xml" value="0.33" />
         <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/app/src/main/res/layout/activity_personal_stats.xml" value="0.20531526836894215" />
+        <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/app/src/main/res/layout/activity_profile.xml" value="0.5" />
+        <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/app/src/main/res/layout/activity_search.xml" value="0.5" />
         <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/code/app/src/main/res/drawable/colorgradient.xml" value="0.3828125" />
         <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/code/app/src/main/res/drawable/edit_text_background.xml" value="0.3828125" />
         <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/code/app/src/main/res/layout/activity_login_test.xml" value="0.25" />
@@ -14,9 +18,7 @@
         <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_personal_stats.xml" value="0.1286231884057971" />
         <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_scan_view.xml" value="0.1286231884057971" />
         <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/toolbar_title_layout.xml" value="0.1286231884057971" />
-        <entry key="..\:/Users/Owner/AndroidStudioProjects/QR-Scape/app/src/main/res/layout/activity_expanded_qr_view.xml" value="0.2796875" />
-        <entry key="..\:/Users/Owner/AndroidStudioProjects/QR-Scape/app/src/main/res/layout/activity_other_profile.xml" value="0.25" />
-        <entry key="..\:/Users/Owner/AndroidStudioProjects/QR-Scape/app/src/main/res/layout/activity_profile.xml" value="0.25" />
+        <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/drawable/bg_round.xml" value="0.16770833333333332" />
         <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_home.xml" value="0.12454710144927536" />
         <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_leaderboard.xml" value="0.30833333333333335" />
         <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_location.xml" value="0.12454710144927536" />
@@ -29,6 +31,10 @@
         <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/leaderboard_item.xml" value="0.30833333333333335" />
         <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/qr_collections_layout.xml" value="0.12454710144927536" />
         <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/qr_list_content.xml" value="0.12454710144927536" />
+        <entry key="app/src/main/res/layout/activity_expanded_qr_view.xml" value="0.2994791666666667" />
+        <entry key="app/src/main/res/layout/activity_profile.xml" value="0.3104166666666667" />
+        <entry key="app/src/main/res/layout/expanded_qr_activity.xml" value="0.3104166666666667" />
+        <entry key="app/src/main/res/layout/leaderboard_item.xml" value="0.30625" />
         <entry key="app/src/main/res/layout/qr_collections_layout.xml" value="0.11851851851851852" />
         <entry key="app/src/main/res/layout/qr_list_content.xml" value="0.27037037037037037" />
       </map>
@@ -36,7 +42,7 @@
   </component>
   <component name="ExternalStorageConfigurationManager" enabled="true" />
   <component name="JavadocGenerationManager">
-    <option name="OUTPUT_DIRECTORY" value="$PROJECT_DIR$/../../../QR-Scape/doc" />
+    <option name="OUTPUT_DIRECTORY" value="$USER_HOME$/QR-Scape/doc" />
   </component>
   <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="11" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
