+++ app/src/main/java/com/example/qr_scape/ProfileActivity.java
@@ -225,8 +225,4 @@ public class ProfileActivity extends AppCompatActivity {
         startActivity(openStatsIntent);
     }
 
-    public void qrCollectionButtonClicked(View view) {
-        startActivity(new Intent(ProfileActivity.this, QRCollectionActivity.class));
-    }
-
 }
