+++ app/src/main/res/layout/activity_profile.xml
@@ -135,17 +135,5 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/profile_code_edittext" />
 
-    <Button
-        android:id="@+id/qrCollectionButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="24dp"
-        android:layout_marginTop="8dp"
-        android:onClick="qrCollectionButtonClicked"
-        android:text="See My QR Codes"
-        android:textColor="@color/white"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/profile_code_edittext" />
-
 
 </androidx.constraintlayout.widget.ConstraintLayout>
